Create Netlogo-Model-Code
The agent-based model's code in Netlogo programming language (for Netlogo simulation software). Some lines of the code are commented out but they can be re-included in order to try different variations of the model. That is, taxation (line 33) may be added to the model and re-adjusted for percentage (originally 10%), and almsgiving percentage may be altered (originally 1/40 - 2.5%, in order to simulate İslamic almsgiving - zakat)